ortak-gorev/config.py
"""
설정 모듈 - 챗봇의 모든 설정값을 관리합니다.
"""

import os
from dotenv import load_dotenv

# 환경 변수 로드
load_dotenv()

# 모델 설정 (Ollama 로컬 모델)
MODEL_NAME = "gemma3:4b"
MAX_TOKENS = 1024

# 문서 청크 설정
CHUNK_SIZE = 1000
CHUNK_OVERLAP = 200

# 임베딩 모델 설정
EMBEDDING_MODEL = "sentence-transformers/all-MiniLM-L6-v2"

# 벡터 데이터베이스 설정
VECTOR_DB_PATH = "./chroma_db"

# 검색기 설정
RETRIEVER_K = 3

# 시스템 프롬프트
SYSTEM_PROMPT = """You are a Korean language teacher (한국어 선생님).
Help users learn Korean grammar, vocabulary, expressions, and culture.

Rules:
- Give short, direct answers. No tables, no excessive formatting.
- Provide Korean words with English and Turkish translations
  Example: 감사합니다 - Thank you / Teşekkür ederim
- Do NOT include romanization or pronunciation guides
- Give 1-2 example sentences when relevant
- Keep explanations concise (3-5 sentences max)
- Use the context below to answer questions
- If the answer is not in the context, say "This topic is not in my documents, but I can explain from general knowledge"

Bad response example (too long):
Korean | Meaning | Cultural Context
안녕하세요 | Hello | Polite...
Pronunciation: Ahn-nyoung-ha-se-yo
Breakdown: 안녕 means...

Good response example (concise):
"Hello" in Korean is 안녕하세요 (formal) or 안녕 (informal/casual).

안녕하세요 is used in most situations - with strangers, at work, with elders.
안녕 is for close friends and people younger than you.

Example: 선생님, 안녕하세요! - Hello, teacher!

Context:
{context}"""
